feat(nes-emu): upgrade Bevy 0.15→0.18 and add emulation pause
Bevy 0.18 migration:
- bevy 0.18, bevy_egui 0.39, egui_plot 0.34
- Handle::weak_from_u128 → uuid_handle! macro
- Material2d/Material2dPlugin → bevy::sprite_render
- ShaderRef → bevy::shader::ShaderRef
- RenderAssetUsages → bevy::asset::RenderAssetUsages
- Image.data is now Option<Vec<u8>>
- EguiContexts::ctx_mut() returns Result
- UI systems moved to EguiPrimaryContextPass schedule
- EventWriter/EventReader replaced with PendingActions resource
- WindowResolution requires (u32, u32) tuple
- Line::new() takes (name, series) in egui_plot 0.34

Pause feature:
- Escape toggles emulation pause
- CPU registers, flags, disassembly, and stack only shown when paused
- Heading shows "PAUSED" indicator when paused
- Accumulator resets on pause to prevent catch-up burst on unpause

+++ nes-emu/src/main.rs
@@ -23,7 +23,7 @@ fn main() {
         .add_plugins(DefaultPlugins.set(WindowPlugin {
             primary_window: Some(Window {
                 title: "NES Emulator".to_string(),
-                resolution: (768.0 + debug_ui::PANEL_WIDTH, 720.0).into(),
+                resolution: ((768.0 + debug_ui::PANEL_WIDTH) as u32, 720u32).into(),
                 present_mode: PresentMode::AutoVsync,
                 // Transparent window prevents OS from painting a white background
                 // before Bevy's first render. ClearColor::BLACK fills it immediately.
